export 'mode' by renaming it to 'AEAD'
It's bad practice to return an unexported type, so export 'mode' and
rename it to something more sensible while we're at it.

New could return a cipher.AEAD interface instead, which is a common
pattern for crypto packages, but returning an interface from a
constructor is also an antipattern which makes it harder to add new
methods to the type later, so just bite the bullet and export it.