Update to version 1.43 of the spec
The H permutation in the reference implementation was mistakenly using
the inverse permutation. The spec has been updated to clarify the
permutation, and reference implementation has been updated to match.

This commit brings us in sync with both the v1.43 spec and the
updated reference implementation. It also updates the test vectors.

Please note that this causes the output of the AEAD to change. Messages
encrypted with commits prior to this will not be able to be decrypted with
this version.

Fixes #1